+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+##[Unreleased]
+#### Fixed
+-Rename assets to wallet, 
+-Hid assets with 0 balance
+
 ## [1.0.0] - 2019-08-23
 #### Added
 - Img url