+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
-##[Unreleased]
+
+## [1.0.1] - 2019-08-27
 #### Added
 - Filter assets on send by asset owner
 
@@ -167,7 +168,8 @@ for further information about branching and tagging conventions.
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/tokend/web-client/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/tokend/web-client/compare/1.0.0-rc.8...1.0.0
 [1.0.0-rc.8]: https://github.com/tokend/web-client/compare/1.0.0-rc.7...1.0.0-rc.8
 [1.0.0-rc.7]: https://github.com/tokend/web-client/compare/1.0.0-rc.6...1.0.0-rc.7